pyproject.toml
[project]
name = "11b-integrate-with-payment"
version = "0.1.0"
description = ""
authors = [
    {name = "OliverRoat",email = "oliver.roat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "stripe (>=12.1.0,<13.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
